no autoclean arg

+++ gobbler.py
@@ -4,7 +4,7 @@ import concurrent.futures
 import logging as log
 import sqlite3
 import time
-from typing import Optional, Tuple
+from typing import Optional, Tuple, List
 
 from mcstatus import MinecraftServer
 from requests import get
@@ -57,7 +57,13 @@ def parse_args():
         required=False,
         default=10000,
         metavar='MAX',
-        help='max amount of entries per server, defaults to 10000'
+        help='max amount of db entries per server, defaults to 10000'
+    )
+    parser.add_argument(
+        '--disable-autoclean',
+        action='store_true',
+        required=False,
+        help='disables the automatic db cleanup'
     )
     return parser.parse_args()
 
@@ -80,7 +86,7 @@ def db_init(file: str) -> sqlite3.Connection:
     return con
 
 
-def parse_servers():
+def parse_servers() -> List[str]:
     out = args.servers
     if args.servers_url is not None:
         log.info('downloading servers file from: ' + args.servers_url)
@@ -92,6 +98,7 @@ def parse_servers():
 def fetch_data(address: str) -> Tuple[str, Optional[int], Optional[int]]:
     log.debug('fetching data for: ' + address)
     try:
+        # TODO: allow non standard ports
         status = MinecraftServer.lookup(address).status()
     except (IOError, ValueError):
         log.debug('error while fetching data for: ' + address)
@@ -104,23 +111,23 @@ def save_data(cursor, status):
     cursor.execute('INSERT OR IGNORE INTO servers VALUES (?)', [address])
     cursor.execute('INSERT INTO data VALUES (?,?,?,?)', (address, int(time.time()), players, ping))
     cursor.execute('SELECT COUNT(*) FROM data WHERE address = (?)', [address])
-    if int(cursor.fetchone()[0]) > args.max_entries:
+    if not args.disable_autoclean and int(cursor.fetchone()[0]) > args.max_entries:
         log.debug('removing oldest 100 entries for ' + address)
         cursor.execute('''DELETE FROM data WHERE address = (?) ORDER BY time ASC LIMIT 100''', [address])
 
 
 if __name__ == '__main__':
     args = parse_args()
-    init_logger(log)
+    init_logger(log, log.DEBUG)
     log.debug('starting gobbler with args: ' + str(args))
     servers = parse_servers()
-    log.info('monitored servers: ' + str(servers))
+    log.info('monitoring ' + str(len(servers)) + ' servers: ' + str(servers))
     db_con = db_init(args.db_file)
     while True:
-        cur = db_con.cursor()
+        db_cur = db_con.cursor()
         with concurrent.futures.ThreadPoolExecutor() as executor:
             for result in executor.map(fetch_data, servers):
-                save_data(cur, result)
+                save_data(db_cur, result)
         db_con.commit()
-        cur.close()
+        db_cur.close()
         time.sleep(args.check_delay)
